Tests/LibWeb/Text/expected/get-bounding-client-rect.txt

 inline   {"x":8,"y":500,"width":784,"height":150,"top":500,"right":792,"bottom":650,"left":8}
{"x":8,"y":650,"width":41.296875,"height":17.46875,"top":650,"right":49.296875,"bottom":667.46875,"left":8}
